refactor without select country
