+++ example/src/App.tsx
@@ -30,9 +30,9 @@ import { TerminationForm } from './Termination';
 import { ContractAmendment } from './ContractAmendment';
 import SyntaxHighlighter from 'react-syntax-highlighter';
 import { docco } from 'react-syntax-highlighter/dist/esm/styles/hljs';
-import { OnboardingForm } from './Onboarding';
-import { OnboardingForm as OnboardingFormWithoutSelectCountry } from './OnboardingWithoutSelectCountryStep';
-import { OnboardingCustomBenefitsForm } from './OnboardingWithCustomBenefits';
+import { OnboardingForm } from './flows/Onboarding/Onboarding';
+import { OnboardingForm as OnboardingFormWithoutSelectCountry } from './flows/Onboarding/OnboardingWithoutSelectCountryStep';
+import { OnboardingCustomBenefitsForm } from './flows/Onboarding/OnboardingWithCustomBenefits';
 import { CostCalculatorWithReplaceableComponents } from './CostCalculatorWithReplaceableComponents';
 import BasicCostCalculatorCode from './BasicCostCalculator?raw';
 import BasicCostCalculatorDefaultValuesCode from './BasicCostCalculatorDefaultValues?raw';
