+++ example/src/flows/Onboarding/OnboardingWithoutSelectCountryStep.tsx
@@ -8,20 +8,49 @@ import {
   EmploymentResponse,
   ContractDetailsFormPayload,
   NormalizedFieldError,
+  $TSFixMe,
 } from '@remoteoss/remote-flows';
-import { useState } from 'react';
-import { ReviewOnboardingStep } from './ReviewOnboardingStep';
+import React, { useState } from 'react';
+import { RemoteFlows } from '../../RemoteFlows';
+import { ReviewOnboardingStep } from '../../ReviewOnboardingStep';
 import { OnboardingAlertStatuses } from './OnboardingAlertStatuses';
-import { RemoteFlows } from './RemoteFlows';
-import { AlertError } from './AlertError';
-import './css/main.css';
+import { AlertError } from '../../AlertError';
+import { ONBOARDING_OPTIONS } from './constants';
+import { getStepTitle, StepsNavigation } from './StepsNavigation';
+import '../../css/main.css';
+
+export const InviteSection = ({
+  title,
+  description,
+  children,
+}: {
+  title: React.ReactNode;
+  description: React.ReactNode;
+  children?: React.ReactNode;
+}) => {
+  return (
+    <div className='rmt-invitation-section'>
+      <h2 className='rmt-invitation-title'>{title}</h2>
+      <p className='rmt-invitation-description'>{description}</p>
+      {children}
+    </div>
+  );
+};
 
 type MultiStepFormProps = {
   onboardingBag: OnboardingRenderProps['onboardingBag'];
   components: OnboardingRenderProps['components'];
 };
 
-const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
+const MultiStepForm = ({ components, onboardingBag }: MultiStepFormProps) => {
+  const {
+    BasicInformationStep,
+    ContractDetailsStep,
+    BenefitsStep,
+    SubmitButton,
+    BackButton,
+    SaveDraftButton,
+  } = components;
   const [errors, setErrors] = useState<{
     apiError: string;
     fieldErrors: NormalizedFieldError[];
@@ -29,17 +58,6 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
     apiError: '',
     fieldErrors: [],
   });
-  const {
-    BasicInformationStep,
-    ContractDetailsStep,
-    BenefitsStep,
-    SubmitButton,
-    BackButton,
-  } = components;
-
-  if (onboardingBag.isLoading) {
-    return <p>Loading...</p>;
-  }
 
   switch (onboardingBag.stepState.currentStep.name) {
     case 'basic_information':
@@ -62,7 +80,10 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
               error: Error;
               fieldErrors: NormalizedFieldError[];
             }) => {
-              setErrors({ apiError: error.message, fieldErrors });
+              setErrors({
+                apiError: error.message,
+                fieldErrors,
+              });
             }}
           />
           <AlertError errors={errors} />
@@ -73,6 +94,23 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
             >
               Create Employment & Continue
             </SubmitButton>
+            <SaveDraftButton
+              onSuccess={() => console.log('saving draft')}
+              onError={({
+                error,
+                fieldErrors,
+              }: {
+                error: Error;
+                fieldErrors: NormalizedFieldError[];
+              }) => {
+                setErrors({
+                  apiError: error.message,
+                  fieldErrors,
+                });
+              }}
+            >
+              Save Draft
+            </SaveDraftButton>
           </div>
         </>
       );
@@ -93,7 +131,12 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
             }: {
               error: Error;
               fieldErrors: NormalizedFieldError[];
-            }) => setErrors({ apiError: error.message, fieldErrors })}
+            }) => {
+              setErrors({
+                apiError: error.message,
+                fieldErrors,
+              });
+            }}
           />
           <AlertError errors={errors} />
           <div className='buttons-container'>
@@ -109,78 +152,30 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
             >
               Continue
             </SubmitButton>
+            <SaveDraftButton
+              onSuccess={() => console.log('saving draft')}
+              onError={({
+                error,
+                fieldErrors,
+              }: {
+                error: Error;
+                fieldErrors: NormalizedFieldError[];
+              }) => {
+                setErrors({
+                  apiError: error.message,
+                  fieldErrors,
+                });
+              }}
+            >
+              Save Draft
+            </SaveDraftButton>
           </div>
         </>
       );
-
     case 'benefits':
       return (
         <div className='benefits-container'>
           <BenefitsStep
-            components={{
-              radio: ({ field, fieldData }) => {
-                const selectedValue = field.value;
-
-                type OptionWithMeta = {
-                  value: string;
-                  label: string;
-                  description?: string;
-                  meta?: { display_cost?: string };
-                };
-
-                return (
-                  <div className='benefit-cards-container'>
-                    {(fieldData.options as OptionWithMeta[] | undefined)?.map(
-                      (option) => {
-                        const isSelected = selectedValue === option.value;
-                        const meta = option.meta || {};
-                        return (
-                          <label
-                            key={option.value}
-                            className={`benefit-card${isSelected ? ' benefit-card--selected' : ''}`}
-                          >
-                            <input
-                              type='radio'
-                              name={field.name}
-                              value={option.value}
-                              checked={isSelected}
-                              onChange={field.onChange}
-                              style={{ display: 'none' }}
-                            />
-                            <div
-                              className='benefit-card__label'
-                              title={option.label}
-                            >
-                              {option.label}
-                            </div>
-                            <div className='benefit-card__summary'>
-                              {option.description || 'Plan summary'}
-                            </div>
-                            <div className='benefit-card__cost'>
-                              {meta.display_cost || ''}
-                            </div>
-                            <button
-                              type='button'
-                              className={`benefit-card__button${isSelected ? ' benefit-card__button--selected' : ''}`}
-                              tabIndex={-1}
-                            >
-                              {isSelected
-                                ? 'Plan Selected!'
-                                : 'Select This Plan'}
-                            </button>
-                            {isSelected && (
-                              <span className='benefit-card__selected-check'>
-                                ✓ Plan Selected!
-                              </span>
-                            )}
-                          </label>
-                        );
-                      },
-                    )}
-                  </div>
-                );
-              },
-            }}
             onSubmit={(payload: BenefitsFormPayload) =>
               console.log('payload', payload)
             }
@@ -190,13 +185,28 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
             }: {
               error: Error;
               fieldErrors: NormalizedFieldError[];
-            }) => setErrors({ apiError: error.message, fieldErrors })}
+            }) => {
+              setErrors({
+                apiError: error.message,
+                fieldErrors,
+              });
+            }}
             onSuccess={(data: SuccessResponse) => console.log('data', data)}
           />
           <AlertError errors={errors} />
           <div className='buttons-container'>
-            <BackButton className='back-button'>Previous Step</BackButton>
-            <SubmitButton className='submit-button'>Continue</SubmitButton>
+            <BackButton
+              className='back-button'
+              onClick={() => setErrors({ apiError: '', fieldErrors: [] })}
+            >
+              Previous Step
+            </BackButton>
+            <SubmitButton
+              onClick={() => setErrors({ apiError: '', fieldErrors: [] })}
+              className='submit-button'
+            >
+              Continue
+            </SubmitButton>
           </div>
         </div>
       );
@@ -205,53 +215,79 @@ const MultiStepForm = ({ onboardingBag, components }: MultiStepFormProps) => {
         <ReviewOnboardingStep
           onboardingBag={onboardingBag}
           components={components}
-          errors={errors}
           setErrors={setErrors}
+          errors={errors}
         />
       );
   }
 };
 
-type OnboardingWithCustomBenefitsProps = {
-  companyId: string;
-  type: 'employee' | 'contractor';
-  employmentId: string;
-  countryCode: string;
-};
+const OnBoardingRender = ({
+  onboardingBag,
+  components,
+}: MultiStepFormProps) => {
+  const currentStepIndex = onboardingBag.stepState.currentStep.index;
+  const stepTitle = getStepTitle(
+    onboardingBag.steps[currentStepIndex],
+    onboardingBag.selectedCountry?.code ?? null,
+  );
 
-const OnboardingWithCustomBenefits = ({
-  type,
-  companyId,
-  employmentId,
-  countryCode,
-}: OnboardingWithCustomBenefitsProps) => {
+  if (onboardingBag.isLoading) {
+    return <p>Loading...</p>;
+  }
+
   return (
-    <RemoteFlows>
-      <OnboardingFlow
-        companyId={companyId}
-        countryCode={countryCode}
-        type={type}
-        render={MultiStepForm}
-        employmentId={employmentId}
-        skipSteps={['select_country']}
+    <>
+      <StepsNavigation
+        steps={onboardingBag.steps}
+        stepState={onboardingBag.stepState}
+        selectedCountry={onboardingBag.selectedCountry}
       />
-    </RemoteFlows>
+
+      <div className='card' style={{ marginBottom: '20px' }}>
+        <h1 className='heading'>{stepTitle}</h1>
+        <MultiStepForm onboardingBag={onboardingBag} components={components} />
+      </div>
+    </>
   );
 };
 
 type OnboardingFormData = {
   companyId: string;
+  countryCode: string;
   type: 'employee' | 'contractor';
   employmentId: string;
-  countryCode: string;
+  externalId?: string;
 };
 
-export const OnboardingCustomBenefitsForm = () => {
+const OnboardingWithProps = ({
+  companyId,
+  type,
+  employmentId,
+  countryCode,
+  externalId,
+}: OnboardingFormData) => (
+  <RemoteFlows>
+    <OnboardingFlow
+      companyId={companyId}
+      type={type}
+      render={OnBoardingRender}
+      employmentId={employmentId}
+      countryCode={countryCode}
+      skipSteps={['select_country']}
+      externalId={externalId}
+      options={ONBOARDING_OPTIONS as $TSFixMe}
+    />
+  </RemoteFlows>
+);
+
+export const OnboardingForm = () => {
   const [formData, setFormData] = useState<OnboardingFormData>({
-    companyId: 'c3c22940-e118-425c-9e31-f2fd4d43c6d8', // use your own company ID
     type: 'employee',
-    countryCode: 'PRT',
     employmentId: '',
+    companyId: 'c3c22940-e118-425c-9e31-f2fd4d43c6d8', // use your own company ID
+    countryCode: 'PRT',
+    externalId: '',
   });
   const [showOnboarding, setShowOnboarding] = useState(false);
 
@@ -261,7 +297,7 @@ export const OnboardingCustomBenefitsForm = () => {
   };
 
   if (showOnboarding) {
-    return <OnboardingWithCustomBenefits {...formData} />;
+    return <OnboardingWithProps {...formData} />;
   }
 
   return (
@@ -282,7 +318,6 @@ export const OnboardingCustomBenefitsForm = () => {
           className='onboarding-form-input'
         />
       </div>
-
       <div className='onboarding-form-group'>
         <label htmlFor='countryCode' className='onboarding-form-label'>
           Country Code:
@@ -299,7 +334,6 @@ export const OnboardingCustomBenefitsForm = () => {
           className='onboarding-form-input'
         />
       </div>
-
       <div className='onboarding-form-group'>
         <label htmlFor='type' className='onboarding-form-label'>
           Type:
@@ -320,7 +354,6 @@ export const OnboardingCustomBenefitsForm = () => {
           <option value='contractor'>Contractor</option>
         </select>
       </div>
-
       <div className='onboarding-form-group'>
         <label htmlFor='employmentId' className='onboarding-form-label'>
           Employment ID:
@@ -336,7 +369,21 @@ export const OnboardingCustomBenefitsForm = () => {
           className='onboarding-form-input'
         />
       </div>
-
+      <div className='onboarding-form-group'>
+        <label htmlFor='externalId' className='onboarding-form-label'>
+          External ID:
+        </label>
+        <input
+          id='externalId'
+          type='text'
+          value={formData.externalId}
+          onChange={(e) =>
+            setFormData((prev) => ({ ...prev, externalId: e.target.value }))
+          }
+          placeholder='Enter External ID'
+          className='onboarding-form-input'
+        />
+      </div>
       <button type='submit' className='onboarding-form-button'>
         Start Onboarding
       </button>
